refactor + add perf metrics
