--- 06_gpu_and_ml/audio-to-text/kyutai/README.md
@@ -1,8 +1,0 @@
-# Kyutai STT
-
-Inspired by [Quillman](https://github.com/modal-labs/quillman).
-
-## File structure
-
-1. FastHTML frontend served by [`src/app.py`](./src/app.py)
-2. Kyutai STT websocket server ([`src/stt.py`](./src/stt.py))
