--- 06_gpu_and_ml/audio-to-text/kyutai/src/common.py
@@ -1,3 +1,0 @@
-from modal import App
-
-app = App(name="kyutai")
